Extract sub components
